requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
faster-whisper==1.0.1
ffmpeg-python==0.2.0
python-multipart==0.0.6
pydantic==2.5.0
